requirements.txt
amqp==5.1.1
anyio==3.6.2
asgiref==3.6.0
async-timeout==4.0.2
attrs==23.1.0
billiard==3.6.4.0
celery==5.2.7
certifi==2023.5.7
charset-normalizer==3.1.0
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
coreapi==2.3.3
coreschema==0.0.4
Django==4.2.1
django-redis==5.2.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
drf-spectacular==0.26.3
drf-spectacular-sidecar==2023.7.1
drf-yasg==1.21.5
gunicorn==20.1.0
idna==3.4
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.2
jsonschema==4.18.0
jsonschema-specifications==2023.6.1
kombu==5.2.4
MarkupSafe==2.1.2
packaging==23.1
prompt-toolkit==3.0.38
pydantic==1.10.7
PyJWT==2.7.0
python-decouple==3.8
pytz==2023.3
PyYAML==6.0
redis==4.5.4
referencing==0.29.1
requests==2.30.0
rpds-py==0.8.8
ruamel.yaml==0.17.26
ruamel.yaml.clib==0.2.7
six==1.16.0
sniffio==1.3.0
sqlparse==0.4.4
starlette==0.26.1
typing_extensions==4.5.0
uritemplate==4.1.1
urllib3==2.0.2
vine==5.0.0
wcwidth==0.2.6
